Fix radio button bug
Changed the radio button for a single checkbox
and fixed the problem with the value storage.
Also add an if statement at the moment of card
creation so it can respond to that value.
Too much repetition going on though, so i have
to refractor that in the future, probably along
with the organization of the code.

Changed the max-width of the cards.