Add required inputs
I couldnt use the required attribute in html, i think
it is because the form is inside a dialog.
So i create a function called dataObservant which is
going to activate when the add button of the field is
pressed. If some data of one input is missing,
preventDefault() is goingto kick in and a message
requiring its completion. Only if all the inputs are
completed, the dataObservant is going to call addBook.

Also, change a little bit in the html to add the p element
to show the message, and styled it in css.